requirements.txt
# ── Web framework ─────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Validation & settings ─────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Database ──────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5

# ── Cache / queue ─────────────────────────────────────────────────────
redis[hiredis]==5.0.4

# ── Rate limiting ─────────────────────────────────────────────────────
slowapi==0.1.9
limits==3.12.0

# ── Crypto / security ─────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.7

# ── HTTP client (webhook delivery, outgoing requests) ─────────────────
httpx==0.27.0

# ── Logging ───────────────────────────────────────────────────────────
structlog==24.1.0
python-json-logger==2.0.7

# ── Observability — Prometheus ────────────────────────────────────────
prometheus-fastapi-instrumentator==6.1.0

# ── Observability — OpenTelemetry (optional; activated via OTEL_ENABLED) ──
# Install these when deploying with OTEL_ENABLED=true.
# Pinned here so `pip install -r requirements.txt` is reproducible.
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-httpx==0.45b0

# ── Utilities ─────────────────────────────────────────────────────────
python-dotenv==1.0.1
numpy==1.26.4
orjson==3.10.3
ulid-py==1.1.0
